Move intake states
