+++ src/main/java/frc/robot/RobotContainer.java
@@ -13,13 +13,13 @@ import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.robot.commands.AprilRunCommand;
 import frc.robot.commands.ArmSetpointCommand;
 import frc.robot.commands.ChargeCircleCommand;
-import frc.robot.commands.ConeGrabState;
-import frc.robot.commands.CubeGrabState;
 import frc.robot.constants.TelemetryConstants.Limelights;
 import frc.robot.subsystems.arm.ArmSubsystem;
 import frc.robot.subsystems.drivetrain.Drivetrain;
 import frc.robot.subsystems.drivetrain.states.DriveState;
 import frc.robot.subsystems.intake.IntakeSubsytstem;
+import frc.robot.subsystems.intake.states.ConeGrabState;
+import frc.robot.subsystems.intake.states.CubeGrabState;
 import frc.robot.utils.XboxController;
 import frc.team4272.controllers.utilities.JoystickAxes;
 import frc.team4272.controllers.utilities.JoystickAxes.DeadzoneMode;
